fix size_t <-> int casting for mmap pows

+++ src/input_chunker.h
@@ -17,7 +17,7 @@ sym_id chunker_in_order(size_t id, size_t ninputs, size_t nsyms);
 size_t rchunker_in_order(sym_id sym, size_t ninputs, size_t nsyms);
 
 void
-type_degree(size_t *rop, acircref ref, const acirc *c, size_t nsyms, input_chunker chunker);
+type_degree(int *rop, acircref ref, const acirc *c, size_t nsyms, input_chunker chunker);
 
 int *
 get_input_syms(const int *inputs, size_t ninputs, reverse_chunker rchunker,
