+++ src/lin/level.c
@@ -105,13 +105,13 @@ level_flatten(int *pows, const level *lvl)
         for (size_t j = 0; j < lvl->c+2; j++) {
             if ((int) lvl->mat[i][j] < 0)
                 return ERR;
-            pows[z++] = lvl->mat[i][j];
+            pows[z++] = (int) lvl->mat[i][j];
         }
     }
     for (size_t i = 0; i < lvl->gamma; i++) {
         if ((int) lvl->vec[i] < 0)
             return ERR;
-        pows[z++] = lvl->vec[i];
+        pows[z++] = (int) lvl->vec[i];
     }
     return OK;
 }
